add ci and count totals
