+++ file/sample_test_file.csv
@@ -5,4 +5,5 @@
 49900000005:F:50300:50300:45320:45320:8478:801325:9900000576:1:E:FE:24/01/2017:OFFICE FOR NATIONAL STATISTICS:::OFFICE FOR NATIONAL STATISTICS:::TOTAL UK ACTIVITY:::C:D:7:0001:S
 49900000006:F:50300:50300:45320:45320:8478:801325:9900000576:1:E:FE:24/01/2017:BOLTS AND RATCHETS:::BOLTS AND RATCHETS:::BOLTS LTD:::C:D:7:0001:S
 49900000007:F:50300:50300:45320:45320:8478:801325:9900000576:1:E:FE:24/01/2017:PC UNIVERSE:::PC UNIVERSE:::PC LTD:::C:D:7:0001:S
+49900000008:F:50300:50300:45320:45320:8478:801325:9900000576:1:E:FE:24/01/2017:RED BRICKS LTD:::RED BRICKS LTD::::::C:D:7:0002:S
 49900000008:F:50300:50300:45320:45320:8478:801325:9900000576:1:E:FE:24/01/2017:RED BRICKS LTD:::RED BRICKS LTD::::::C:D:7:0001:S
