+++ go.mod
@@ -6,10 +6,10 @@ require (
 	cloud.google.com/go/pubsub v1.5.0
 	github.com/google/wire v0.4.0
 	github.com/gorilla/mux v1.7.4
+	github.com/magiconair/properties v1.8.1
 	github.com/sirupsen/logrus v1.6.0
 	github.com/spf13/viper v1.7.0
-	github.com/stretchr/testify v1.4.0
+	github.com/stretchr/testify v1.6.0
 	google.golang.org/api v0.28.0
 	google.golang.org/grpc v1.29.1
-	github.com/stretchr/testify v1.6.0
 )
